Fix test not working
