+++ app/src/androidTest/java/ch/epfl/sdp/drone3d/ui/map/MissionInProgressActivityTest.kt
@@ -46,13 +46,11 @@ class MissionInProgressActivityTest {
     @BindValue
     val droneService: DroneService = DroneInstanceMock.mockService()
 
-    private val statusLiveData = MutableLiveData<DroneData.DroneStatus>()
-
     init {
         `when`(droneService.getData().isConnected()).thenReturn(MutableLiveData(true))
         `when`(droneService.getData().getPosition()).thenReturn(MutableLiveData())
         `when`(droneService.getData().getHomeLocation()).thenReturn(MutableLiveData())
-        `when`(droneService.getData().getDroneStatus()).thenReturn(statusLiveData)
+        `when`(droneService.getData().getDroneStatus()).thenReturn(MutableLiveData())
         `when`(droneService.getData().isConnected()).thenReturn(MutableLiveData())
         `when`(droneService.getData().getVideoStreamUri()).thenReturn(MutableLiveData())
         `when`(droneService.getData().getMission()).thenReturn(MutableLiveData())
@@ -91,23 +89,5 @@ class MissionInProgressActivityTest {
         }
 
         ToastMatcher.onToast(activity.get(), R.string.lost_connection_message)
-    }
-
-    @Test
-    fun buttonsBecomeVisibleWhenMissionIsExecuted() {
-        statusLiveData.postValue(DroneData.DroneStatus.IDLE)
-
-        onView(withId(R.id.backToHomeButton))
-                .check(matches(withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
-        onView(withId(R.id.backToUserButton))
-                .check(matches(withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
-
-
-        statusLiveData.postValue(DroneData.DroneStatus.EXECUTING_MISSION)
-
-        onView(withId(R.id.backToHomeButton))
-                .check(matches(withEffectiveVisibility(ViewMatchers.Visibility.VISIBLE)))
-        onView(withId(R.id.backToUserButton))
-                .check(matches(withEffectiveVisibility(ViewMatchers.Visibility.VISIBLE)))
     }
 }
